Project first segment
I’m uploading this as a back-up more than anything. The game functions
at this point, although there are only four “rooms” and not a lot of
story to go along with it. However, the user’s character can take and
drop items that they find in rooms, and move from room to room. 6/22/14.